+++ src/util/utils.c
@@ -1,9 +1,8 @@
-#include "utils.h"
-#include<stdio.h>
+#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <errno.h>
-#include <limits.h>
+
+#include "utils.h"
 
 int calcCheckSum(char *data, int size) {
 	int checksum = 0;
@@ -21,7 +20,7 @@ char *IntToString(int data) {
 	int size = (data / 10) + 1;
 	char *string = (char *) malloc(sizeof(char) * (size + 1));
 	sprintf(string, "%d", data);
-    string[size] = '\0';
+	string[size] = '\0';
 	return string;
 }
 
@@ -30,7 +29,7 @@ char *padLeft(char *data, char c, int lenght) {
 	int size = (int) strlen(data);
 	if (size < lenght) {
 		string = (char *) malloc(sizeof(char) * (lenght + 1));
-        memset(string,'\0',lenght + 1);
+		memset(string, '\0', lenght + 1);
 		for (int i = lenght - 1; i >= lenght - size; i--) {
 			string[i] = data[i - lenght + size];
 		}
@@ -47,7 +46,7 @@ char *substring(const char *data, int start, int end) {
 	for (int i = start; i < end; i++) {
 		string[i - start] = data[i];
 	}
-    string[lenght] = '\0';
+	string[lenght] = '\0';
 	return string;
 }
 
@@ -58,8 +57,8 @@ char *substring(const char *data, int start, int end) {
  * @return char* concat string
  */
 char *concat(char *s1, char *s2) {
-	int l1 = strlen(s1);
-	int l2 = strlen(s2);
+	int l1 = (int) strlen(s1);
+	int l2 = (int) strlen(s2);
 	char *string = (char *) malloc(sizeof(char) * (l1 + l2 + 1));
 	for (int i = 0; i < l1; i++) {
 		string[i] = s1[i];
@@ -67,7 +66,7 @@ char *concat(char *s1, char *s2) {
 	for (int i = l1; i < l1 + l2; i++) {
 		string[i] = s2[i - l1];
 	}
-    string[l1 + l2] = '\0';
+	string[l1 + l2] = '\0';
 	return string;
 }
 
@@ -79,17 +78,17 @@ char *concat(char *s1, char *s2) {
 char *copy(char *s) {
 	int lenght = (int) strlen(s);
 	char *string = (char *) malloc(sizeof(char) * (lenght + 1));
-    string[lenght] = '\0';
+	string[lenght] = '\0';
 	for (int i = 0; i < lenght; i++) {
 		string[i] = s[i];
 	}
 	return string;
 }
 
-int isNumber(char * string){
-    for(int i = 0; i < strlen(string); i++){
-        if (string[i] > '9' || string[i] < '0')
-            return 0;
-    }
-    return 1;
+int isNumber(char *string) {
+	for (int i = 0; i < strlen(string); i++) {
+		if (string[i] > '9' || string[i] < '0')
+			return 0;
+	}
+	return 1;
 }
